refactor(app): split standalone chat internals
Break the standalone chat implementation into focused helper modules, render components, and local state hooks while preserving the existing public import path and runtime behavior.

Share chat message and content block types with the conversation hook to remove duplicated type definitions.
